+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+v0.20.0 - _October 3, 2017_
+    * Add `zeroEx.token.getLogsAsync` (#178)
+    * Add `zeroEx.exchange.getLogsAsync` (#178)
+
 v0.19.0 - _September 29, 2017_
     * Made order validation optional  (#172)
     * Added Ropsten testnet support (#173)